reuse existing breakdown property map
